LayoutTests/fullscreen/full-screen-iframe-with-max-width-height-expected.txt


EVENT(webkitfullscreenchange) TEST(frame.getBoundingClientRect().right == document.body.getBoundingClientRect().right && frame.getBoundingClientRect().bottom == document.body.getBoundingClientRect().bottom) OK
END OF TEST
